#80 feat: assembler 제거 -> service 레이어에 역할 위임
